we have skinny routes and fat models now
